+++ changelog.md
@@ -3,6 +3,8 @@
 ## Upcoming
 * Sponges
   * Wet Sponges now dry when exposed to sun and Sponges now become wet when exposed to rain 
+* Graves
+  * Graves now save the death message of the player and can be seen when right-clicked
 * Tiredness
   * When tired, random sounds will be played to the player
   * Increased stamina consumed and lowered stamina regenerated with the Tired effect